requirements.txt
anthropic>=0.39.0
openai>=1.50.0
reportlab>=4.1
Pillow>=10.0
python-dotenv>=1.0
